TST: Removed pytest_jwst plugin because it is not really needed.
