+++ jwst/conftest.py
@@ -204,3 +204,25 @@ def log_watcher(monkeypatch):
         return watcher
 
     return _log_watcher
+
+
+def pytest_report_header(config):
+    """
+    Add CRDS_CONTEXT to pytest report header.
+
+    Parameters
+    ----------
+    config : pytest.config.Config
+        Pytest configuration object.
+
+    Returns
+    -------
+    str
+        Report header string with CRDS context information.
+    """
+    if config.getoption("report_crds_context"):
+        from stpipe.crds_client import get_context_used
+
+        return f"crds_context: {get_context_used('jwst')}"
+    else:
+        return []
