--- pytest_jwst/__init__.py
@@ -1,1 +1,0 @@
-"""Pytest plugins for JWST."""
